+++ build-tools/webpack/webpack.build.config.js
@@ -2,7 +2,6 @@ const common = require('./webpack.common.config');
 const merge = require('webpack-merge');
 const webpack = require('webpack');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const path = require('path');
 const componentEntries = require('./webpack.componentEntries.config');
 
 module.exports = merge(componentEntries, common, {
