+++ package.json
@@ -5,8 +5,9 @@
   "main": "index.js",
   "scripts": {
     "app:build": "node_modules/.bin/webpack --config ./build-tools/webpack/webpack.build.config.js",
+    "app:clear": "rimraf dist/*",
     "app:start": "node_modules/.bin/webpack-dev-server --hot --config ./build-tools/webpack/webpack.dev.config.js",
-    "jigs:build": "node_modules/.bin/tsc",
+    "jigs:build": "npm run app:clear && npm run app:build",
     "jigs:watch": "",
     "test": "node_modules/.bin/karma start testing/karma.conf.js"
   },
@@ -43,6 +44,7 @@
     "karma-webpack": "2.0.13",
     "mocha": "5.0.4",
     "node-sass": "4.7.2",
+    "rimraf": "2.6.2",
     "sass-loader": "6.0.7",
     "style-loader": "0.20.2",
     "stylelint": "9.1.1",
